meliusltexx: Define build fingerprint and description
This is just to work around an issue in Android which prevents booting
when the fingerprint is longer than 92 chars. In 9.0.0_r16 Google
extended the build ID by 3 chars, thus making the auto-generated
fingerprint during lineage building too long. This will be overridden
by our custom init anyway, according to the device variant.

Notice to Google: What's the point of allowing more than 92 chars in
ro. property values during build, when it's causing boot failure
when a property exceeds it?

Change-Id: Ibb9d2cafc9533d120e212f8aaf2c3e554b9248b2